hw1/task2-1/requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bitsandbytes==0.45.3
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==3.4.1
diffusers==0.32.2
dill==0.3.8
einops==0.8.1
filelock @ file:///croot/filelock_1700591183607/work
flash_attn==2.7.4.post1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gmpy2 @ file:///croot/gmpy2_1738085463648/work
huggingface-hub==0.29.3
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.6.1
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.1
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.2.0
multiprocess==0.70.16
networkx @ file:///croot/networkx_1737039604450/work
nltk==3.9.1
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp310-cp310-linux_x86_64.whl#sha256=120568f3fd675f59e4cb6de79a5c193b4067d90878d8bab2040fda8e3a2df6fa
packaging==24.2
pandas==2.2.3
peft==0.15.0
pillow @ file:///croot/pillow_1738010226202/work
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
pyarrow==19.0.1
pycocotools==2.0.8
pyparsing==3.2.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rouge==1.0.1
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.2
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
urllib3 @ file:///croot/urllib3_1737133630106/work
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
